requirements.txt
Flask==3.0.3
psutil==6.0.0
speedtest-cli==2.1.3
gunicorn==22.0.0
